chore(test): replace bootstrap by mocha hook in functional tests.
