+++ test/functional/3dtiles_basic.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('3dtiles_basic', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/3dtiles_basic.html`, this.fullTitle());
+        result = await loadExample('examples/3dtiles_basic.html', this.fullTitle());
     });
 
     it('should run', async () => {
