+++ test/functional/effects_postprocessing.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('effects_postprocessing', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/effects_postprocessing.html`, this.fullTitle());
+        result = await loadExample('examples/effects_postprocessing.html', this.fullTitle());
     });
 
     it('should run', async () => {
