+++ test/functional/misc_colorlayer_visibility.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('misc_colorlayer_visibility', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/misc_colorlayer_visibility.html`, this.fullTitle());
+        result = await loadExample('examples/misc_colorlayer_visibility.html', this.fullTitle());
     });
 
     it('should run', async () => {
