+++ test/functional/oriented_images.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('misc_georeferenced_images', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/misc_georeferenced_images.html`, this.fullTitle());
+        result = await loadExample('examples/misc_georeferenced_images.html', this.fullTitle());
     });
 
     it('should run', async () => {
