+++ test/functional/source_file_gpx_3d.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('source_file_gpx_3d', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/source_file_gpx_3d.html`, this.fullTitle());
+        result = await loadExample('examples/source_file_gpx_3d.html', this.fullTitle());
     });
 
     it('should run', async () => {
