+++ test/functional/source_stream_wfs_25d.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('source_stream_wfs_25d', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/source_stream_wfs_25d.html`, this.fullTitle());
+        result = await loadExample('examples/source_stream_wfs_25d.html', this.fullTitle());
     });
 
     it('should run', async () => {
