+++ test/functional/vector_tile_raster_3d.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('vector_tile_raster_3d', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/vector_tile_raster_3d.html`, this.fullTitle());
+        result = await loadExample('examples/vector_tile_raster_3d.html', this.fullTitle());
     });
 
     it('should run', async () => {
