+++ test/functional/view_3d_map.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('view_3d_map', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/view_3d_map.html`, this.fullTitle());
+        result = await loadExample('examples/view_3d_map.html', this.fullTitle());
     });
 
     it('should run', async () => {
