+++ test/functional/view_multi_25d.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 describe('view_multi_25d', function _() {
     let result;
     before(async () => {
-        result = await loadExample(`http://localhost:${itownsPort}/examples/view_multi_25d.html`, this.fullTitle());
+        result = await loadExample('examples/view_multi_25d.html', this.fullTitle());
     });
 
     it('should run', async () => {
